refactor: separate components into respective directories
