+++ src/components/filter/Filter.css
@@ -36,3 +36,37 @@
 .filter-container button:active {
   background: #0e19f388;
 }
+
+@media (max-width: 768px) {
+  .filter-container {
+    padding: 12px;
+  }
+  .filter-container h3 {
+    font-size: 1rem;
+  }
+
+  .filter-container label {
+    margin-bottom: 6px;
+  }
+
+  .filter-container button {
+    padding: 6px 12px;
+  }
+}
+
+@media (max-width: 480px) {
+  .filter-container {
+    padding: 8px;
+  }
+  .filter-container h3 {
+    font-size: 0.9rem;
+  }
+
+  .filter-container label {
+    margin-bottom: 4px;
+  }
+
+  .filter-container button {
+    padding: 4px 8px;
+  }
+}
